+++ Source/Objects/Hardware/Hardware Base Classes/ORCrate.m
@@ -71,10 +71,10 @@ NSString* ORCrateModelLockMovementChanged   = @"ORCrateModelLockMovementChanged"
 	[aConnector release];
 }
 
-- (void) setUniqueIdNumber:(unsigned long)anIdNumber
+- (void) setUniqueIdNumber:(uint32_t)anIdNumber
 {
 	[super setUniqueIdNumber:anIdNumber];
-	[self setCrateNumber:anIdNumber-1];
+	[self setCrateNumber:(int)anIdNumber-1];
 }
 
 - (NSString*) fullID
@@ -216,7 +216,7 @@ NSString* ORCrateModelLockMovementChanged   = @"ORCrateModelLockMovementChanged"
     }
 }
 
-- (int)tag
+- (NSUInteger)tag
 {
     return crateNumber;
 }
@@ -291,9 +291,9 @@ NSString* ORCrateModelLockMovementChanged   = @"ORCrateModelLockMovementChanged"
 	[aCard positionConnector:aConnector];
 }
 
-- (unsigned long) requestGTID
+- (uint32_t) requestGTID
 {
-    unsigned long aGTID = 0;
+    uint32_t aGTID = 0;
     NSArray* gtidGenerators = [self collectObjectsConformingTo:@protocol(ORGTIDGenerator)];
     if([gtidGenerators count]){
         id gtidGenerator = [gtidGenerators objectAtIndex:0];
